+++ standard/packages/timegrid/src/components/TimeGridLayout.tsx
@@ -195,13 +195,15 @@ export class TimeGridLayout extends BaseComponent<TimeGridLayoutProps> {
   /*
   Captures current values
   */
-  private handleTimeScrollEnd = () => {
-    // record, but only if not forPrint, which could give bogus values in the case of
-    // TimeGridLayoutPannable, which kills y-scrolling, but retains x-scrolling,
-    // which reports as a 0 y-scroll.
-    if (!this.props.forPrint) {
-      this.scrollState.y = this.timeScrollerRef.current.y
-      this.scrollState.time = undefined
+  private handleTimeScrollEnd = ({ y, isUser }: { y: number, isUser: boolean }) => {
+    if (isUser) {
+      // record, but only if not forPrint, which could give bogus values in the case of
+      // TimeGridLayoutPannable, which kills y-scrolling, but retains x-scrolling,
+      // which reports as a 0 y-scroll.
+      if (!this.props.forPrint) {
+        this.scrollState.y = y
+        this.scrollState.time = undefined
+      }
     }
   }